# perf: speed up creating maps of subdocuments