--- a/lib/document.js
+++ b/lib/document.js
@@ -2226,7 +2226,7 @@ Document.prototype.isModified = function(paths, modifiedPaths) {
       return false;
     }
     for (const path of paths) {
-      if (directModifiedPathsObj.hasOwnProperty(path)) {
+      if (Object.prototype.hasOwnProperty.call(directModifiedPathsObj, path)) {
         return true;
       }
     }