--- a/lib/types/map.js
+++ b/lib/types/map.js
@@ -141,12 +141,12 @@ class MongooseMap extends Map {
     super.set(key, value);
 
     if (value != null && value.$isSingleNested) {
-      value.$basePath = this.$__path + '.' + key;
+      value.$basePath = fullPath;
     }
 
     const parent = this.$__parent;
     if (parent != null && parent.$__ != null && !deepEqual(value, priorVal)) {
-      parent.markModified(this.$__path + '.' + key);
+      parent.markModified(fullPath);
     }
   }