--- a/lib/types/map.js
+++ b/lib/types/map.js
@@ -96,9 +96,9 @@ class MongooseMap extends Map {
       return;
     }
 
-    const fullPath = this.$__path + '.' + key;
+    let _fullPath;
     const populated = this.$__parent != null && this.$__parent.$__ ?
-      this.$__parent.$populated(fullPath, true) || this.$__parent.$populated(this.$__path, true) :
+      this.$__parent.$populated(this.$__path, true) :
       null;
     const priorVal = this.get(key);
 
@@ -127,11 +127,19 @@ class MongooseMap extends Map {
       }
     } else {
       try {
-        value = this.$__schemaType.
-          applySetters(value, this.$__parent, false, this.get(key), { path: fullPath });
+        const options = this.$__schemaType.$isMongooseDocumentArray ?
+          { path: fullPath.call(this) } :
+          null;
+        value = this.$__schemaType.applySetters(
+          value,
+          this.$__parent,
+          false,
+          this.get(key),
+          options
+        );
       } catch (error) {
         if (this.$__parent != null && this.$__parent.$__ != null) {
-          this.$__parent.invalidate(fullPath, error);
+          this.$__parent.invalidate(fullPath.call(this), error);
           return;
         }
         throw error;
@@ -141,12 +149,22 @@ class MongooseMap extends Map {
     super.set(key, value);
 
     if (value != null && value.$isSingleNested) {
-      value.$basePath = fullPath;
+      value.$basePath = fullPath.call(this);
     }
 
     const parent = this.$__parent;
     if (parent != null && parent.$__ != null && !deepEqual(value, priorVal)) {
-      parent.markModified(fullPath);
+      parent.markModified(fullPath.call(this));
+    }
+
+    // Delay calculating full path unless absolutely necessary, because string
+    // concatenation is a bottleneck re: #13171
+    function fullPath() {
+      if (_fullPath) {
+        return _fullPath;
+      }
+      _fullPath = this.$__path + '.' + key;
+      return _fullPath;
     }
   }