--- a/lib/types/map.js
+++ b/lib/types/map.js
@@ -97,8 +97,9 @@ class MongooseMap extends Map {
     }
 
     let _fullPath;
-    const populated = this.$__parent != null && this.$__parent.$__ ?
-      this.$__parent.$populated(this.$__path, true) :
+    const parent = this.$__parent;
+    const populated = parent != null && parent.$__ && parent.$__.populated ?
+      parent.$populated(fullPath.call(this), true) || parent.$populated(this.$__path, true) :
       null;
     const priorVal = this.get(key);
 
@@ -127,7 +128,7 @@ class MongooseMap extends Map {
       }
     } else {
       try {
-        const options = this.$__schemaType.$isMongooseDocumentArray ?
+        const options = this.$__schemaType.$isMongooseDocumentArray || this.$__schemaType.$isSingleNested ?
           { path: fullPath.call(this) } :
           null;
         value = this.$__schemaType.applySetters(
@@ -148,11 +149,6 @@ class MongooseMap extends Map {
 
     super.set(key, value);
 
-    if (value != null && value.$isSingleNested) {
-      value.$basePath = fullPath.call(this);
-    }
-
-    const parent = this.$__parent;
     if (parent != null && parent.$__ != null && !deepEqual(value, priorVal)) {
       parent.markModified(fullPath.call(this));
     }