<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory.

If you're making a change to documentation, do **not** modify a `.html` file directly. Instead, find the corresponding `.pug` file or test case in the `test/docs` directory. -->

**Summary**

Re: #13191, #13271.

Benchmark from #13191 with 2000 documents, before this change:

```
Starting the process of mapping all minis and timing it
Iterating through all minis: 5.661ms
Sorting all minis: 1.368ms
Creating and populating Map of Minis: 0.4ms
Saving Map of Minis to MongoDB using Mongoose: 1.684s
Map Process: 1.692s
Saved Minis Map successfully!
```

After:

```
Starting the process of mapping all minis and timing it
Iterating through all minis: 6.554ms
Sorting all minis: 1.872ms
Creating and populating Map of Minis: 0.555ms
Saving Map of Minis to MongoDB using Mongoose: 315.028ms
Map Process: 324.444ms
Saved Minis Map successfully!
```

I would love to make this faster and will work on it some more, as well as adding a benchmark for this, but figured I'd add a PR first.

<!-- Explain the **motivation** for making this change. What problem does the pull request solve? -->

**Examples**

<!-- If this code fixes a bug or adds a new feature, provide an example demonstrating the change, unless you added a test. -->